Make global lookup cache introspector class public * Fixes #3187
